@@ -5,7 +5,11 @@ import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
+import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -13,12 +17,15 @@ import org.springframework.context.annotation.FilterType;
 import org.springframework.web.client.RestTemplate;
 
 @SpringBootApplication
-@ComponentScan(basePackages={ "io.mosip.commons.khazana.*" ,"com.reencryptutility.*"}, excludeFilters = {
+@ComponentScan(basePackages={ "io.mosip.commons.khazana.*" ,"com.reencryptutility.*", "io.mosip.*", "io.mosip.kernel.auth.defaultadapter"},
+        excludeFilters = {
         @ComponentScan.Filter(type = FilterType.ASPECTJ,
                 pattern = {
                         "io.mosip.commons.khazana.impl.PosixAdapter",
                         "io.mosip.commons.khazana.util.*"
                 }) })
+@EnableAutoConfiguration(exclude = { DataSourceAutoConfiguration.class,
+        DataSourceTransactionManagerAutoConfiguration.class, HibernateJpaAutoConfiguration.class})
 
 /**
  * The Class ReEncryptUtilityApplication.
